# Membership: user_can_view_post: Memoize results