--- a/projects/plugins/jetpack/changelog/update-user-can-view-post-cache
+++ b/projects/plugins/jetpack/changelog/update-user-can-view-post-cache
@@ -1,0 +1,4 @@
+Significance: patch
+Type: other
+
+Membership: user_can_view_post: Memoize results