--- a/projects/plugins/jetpack/modules/memberships/class-jetpack-memberships.php
+++ b/projects/plugins/jetpack/modules/memberships/class-jetpack-memberships.php
@@ -459,7 +459,7 @@ class Jetpack_Memberships {
 	 */
 	public static function user_can_view_post() {
 		$user_id   = get_current_user_id();
-		$post_id   = get_the_ID();
+		$post_id   = get_the_ID() || 0;
 		$cache_key = sprintf( '%d_%d', $user_id, $post_id );
 
 		if ( isset( self::$user_can_view_post_cache[ $cache_key ] ) ) {